+++ test/sql/function/list/list_has_any_and_has_all.test
@@ -175,7 +175,7 @@ select list_has_all([1, 2], 1);
 statement error
 select list_has_any([[1,2], [2,4]], ['abc', 'def']);
 ----
-Binder Error: list_has_any: cannot compare lists of different types: 'INTEGER[]' and 'VARCHAR'
+Binder Error: 'list_has_any' cannot compare lists of different types: 'INTEGER[]' and 'VARCHAR'
 
 # Test Function in Function
